RENDEZVOUS-90:moved to github
NOTE: this commit was intentionally modified to conflict with old
histories that contain the 2 previous direct parents of this commit
(40160f9af8d and e42aed8b98), because they add/remove files and make the
repository 60Mb larger than what it should be. These 2 commits have been
removed and the conflict prevents someone to bring them back with an
automatic successful merge.